removed tensorflow NOTMAD and tensorflow dependencies
