+++ contextualized/dags/callbacks.py
@@ -4,8 +4,7 @@ from pytorch_lightning.callbacks import Callback
 from pytorch_lightning.callbacks import EarlyStopping
 
 # local imports
-import contextualized.dags.torch_notmad.graph_utils
-from contextualized.dags.torch_notmad.torch_utils import DAG_loss
+from contextualized.dags.torch_utils import DAG_loss
 
 early_stopping = EarlyStopping("val_loss")
