+++ setup.py
@@ -5,7 +5,6 @@ setuptools.setup(name='contextualized',
         'contextualized',
         'contextualized.regression',
         'contextualized.dags',
-        'contextualized.dags.notmad_helpers',
         'contextualized.easy',
     ],
     version='0.1.1',
@@ -16,12 +15,6 @@ setuptools.setup(name='contextualized',
         'tqdm',
         'scikit-learn',
         'python-igraph',
-        'matplotlib',
-        'pandas',
-        'umap-learn',
-        'interpret',
-        'tensorflow>=2.4.0',
-        'tensorflow-addons',
         'numpy>=1.19.2',
         'ipywidgets',
         'torchvision',
