+++ tests.py
@@ -1,7 +1,6 @@
 import unittest
 from contextualized.regression.tests import *
 from contextualized.dags.tests import *
-from contextualized.dags.torch_notmad.tests import *
 
 
 if __name__ == '__main__':
